+++ ThirdPartyAdapters/inmobi/inmobi/src/main/java/com/google/ads/mediation/inmobi/InMobiExtrasBuilder.java
@@ -5,10 +5,11 @@ import android.content.SharedPreferences;
 import android.os.Bundle;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import com.google.ads.mediation.common.AgeRestrictedTreatmentUtils;
+import com.google.android.gms.ads.AgeRestrictedTreatment;
 import com.google.android.gms.ads.MobileAds;
 import com.google.android.gms.ads.RequestConfiguration;
 import com.inmobi.compliance.InMobiPrivacyCompliance;
-
 import java.util.HashMap;
 
 public class InMobiExtrasBuilder {
@@ -44,10 +45,16 @@ public class InMobiExtrasBuilder {
     map.put(THIRD_PARTY_VERSION, MobileAds.getVersion().toString());
     // If the COPPA value isn't specified by the publisher, InMobi SDK expects the default value to
     // be `0`.
+    boolean isAgeRestrictedTreatmentChild =
+        AgeRestrictedTreatmentUtils.runtimeGmaSdkSupportsChildAgeRestrictedTreatment()
+            && MobileAds.getRequestConfiguration().getAgeRestrictedTreatment()
+                == AgeRestrictedTreatment.CHILD;
     if (MobileAds.getRequestConfiguration().getTagForChildDirectedTreatment()
-        == RequestConfiguration.TAG_FOR_CHILD_DIRECTED_TREATMENT_TRUE) {
+            == RequestConfiguration.TAG_FOR_CHILD_DIRECTED_TREATMENT_TRUE
+        || isAgeRestrictedTreatmentChild) {
       map.put(COPPA, "1");
     } else {
+      // TODO (b/527112815): Update the default TFCD and TFUA behavior and get 3p review
       map.put(COPPA, "0");
     }